写入布尔值时，强制转为字节，否则会写入4字节整数，二进制读写全部测试通过

+++ NewLife.Core/Serialization/Binary/BinaryGeneral.cs
@@ -40,7 +40,7 @@ namespace NewLife.Serialization
             switch (Type.GetTypeCode(type))
             {
                 case TypeCode.Boolean:
-                    Host.Write((Boolean)value ? 1 : 0);
+                    Host.Write((Byte)((Boolean)value ? 1 : 0));
                     return true;
                 case TypeCode.Byte:
                 case TypeCode.SByte:
